Use trpc-cli class groups for OS scripts
