+++ apps/os/package.json
@@ -95,7 +95,7 @@
     "miniflare": "catalog:cloudflare",
     "openai": "^6.36.0",
     "tinyexec": "^1.0.2",
-    "trpc-cli": "0.15.0",
+    "trpc-cli": "https://pkg.pr.new/mmkal/trpc-cli@211",
     "tsx": "^4.20.3",
     "typescript": "^5.9.3",
     "vite": "^8.0.0",
