+++ apps/semaphore/package.json
@@ -47,7 +47,7 @@
     "react-dom": "^19.2.0",
     "sqlfu": "0.0.3-14",
     "tailwindcss": "^4.1.18",
-    "trpc-cli": "0.15.0",
+    "trpc-cli": "https://pkg.pr.new/mmkal/trpc-cli@211",
     "zod": "^4.3.6"
   },
   "devDependencies": {
