+++ knip.ts
@@ -27,10 +27,8 @@ function makeOsCloudflareAppWorkspace(workerEnvShim: string): WorkspaceConfig {
       "e2e/tui-test/tui-test.config.ts",
       "e2e/tui-test/run.ts",
       // Local operational commands mounted by scripts/cli.ts.
-      "scripts/artifacts.ts",
       "scripts/cli.ts",
       "scripts/dev.ts",
-      "scripts/itx.ts",
       "sqlfu.config.ts",
       "src/durable-objects/codemode-session.vitest.config.ts",
       "src/durable-objects/codemode-session-test-entry.ts",
