+++ packages/iterate/package.json
@@ -59,7 +59,7 @@
     "@orpc/server": "1.13.14",
     "better-auth": "1.4.3",
     "react": "^19.2.0",
-    "trpc-cli": "0.15.0",
+    "trpc-cli": "https://pkg.pr.new/mmkal/trpc-cli@211",
     "yaml": "^2.8.3",
     "zod": "4.1.12"
   },
